+++ backend/cwa_lib/routers/misc.py
@@ -5,7 +5,6 @@ from common.sql_db_async import async_get_session
 from common.sql_models import User
 from cwa_lib.pydantic_schemas.user import ChangePasswordRequest
 from cwa_lib.app import current_active_user, password_helper
-from cwa_lib.pages.su_server_status import SuServerStatusPage
 
 
 router__misc = APIRouter()